+++ crates/node/src/command.rs
@@ -7,8 +7,8 @@ use pallet_starknet::utils;
 use sc_cli::{ChainSpec, RpcMethods, RuntimeVersion, SubstrateCli};
 
 use crate::benchmarking::{inherent_benchmark_data, RemarkBuilder};
-use crate::cli::{Cli, Subcommand, Testnet};
-use crate::{chain_spec, constants, service};
+use crate::cli::{Cli, ExtendedRunCmd, SetupCmd, Subcommand, Testnet};
+use crate::{chain_spec, configs, constants, service};
 impl SubstrateCli for Cli {
     fn impl_name() -> String {
         "Madara Node".into()
@@ -40,11 +40,11 @@ impl SubstrateCli for Cli {
                 let enable_manual_seal = self.sealing.map(|_| true);
                 Box::new(chain_spec::development_config(
                     enable_manual_seal,
-                    self.run.madara_path.clone().expect("Failed retrieving madara_path"),
+                    self.madara_path.clone().expect("`madara_path` expected to be set with clap default value"),
                 )?)
             }
             "" | "local" | "madara-local" => Box::new(chain_spec::local_testnet_config(
-                self.run.madara_path.clone().expect("Failed retrieving madara_path"),
+                self.madara_path.clone().expect("`madara_path` expected to be set with clap default value"),
             )?),
             path => Box::new(chain_spec::ChainSpec::from_json_file(std::path::PathBuf::from(path))?),
         })
@@ -55,110 +55,147 @@ impl SubstrateCli for Cli {
     }
 }
 
-/// Parse and run command line arguments
-pub fn run() -> sc_cli::Result<()> {
-    let mut cli = Cli::from_args();
+fn get_madara_path_string(madara_path: &Option<PathBuf>) -> String {
+    madara_path
+        .clone()
+        .expect("`madara_path` expected to be set with clap default value")
+        .into_os_string()
+        .into_string()
+        .expect("Failed to convert `madara_path` to string")
+}
 
-    // alias madara_path <> base_path
-    // TODO also alias tmp (tmp generates random base_paths that are not specified within
-    // the command)
-    let madara_path = match (cli.run.madara_path.clone(), cli.run.run_cmd.shared_params.base_path.clone()) {
-        (Some(madara_path), _) => {
-            cli.run.run_cmd.shared_params.base_path = Some(madara_path.clone());
-            madara_path.to_str().unwrap().to_string()
-        }
-        (_, Some(base_path)) => {
-            cli.run.madara_path = Some(base_path.clone());
-            base_path.to_str().unwrap().to_string()
-        }
-        _ => {
-            let home_path = std::env::var("HOME").unwrap_or(std::env::var("USERPROFILE").unwrap_or(".".into()));
-            let path = format!("{}/.madara", home_path);
-            cli.run.run_cmd.shared_params.base_path = Some((path.clone()).into());
-            cli.run.madara_path = Some((path.clone()).into());
-            path
-        }
-    };
+fn set_dev_environment(cmd: &mut ExtendedRunCmd) {
+    // create a reproducible dev environment
+    cmd.run_cmd.shared_params.dev = false;
+    cmd.run_cmd.shared_params.chain = Some("dev".to_string());
 
-    if let Some(genesis_url) = cli.run.genesis_url.clone() {
-        // can't copy extra genesis-assets atm
-        // we can reuse #982 to create the standard to fetch relevant files
-        utils::fetch_from_url(genesis_url, madara_path.clone() + "/configs/genesis-assets")?;
-    } else {
-        // TODO confirm with the CI that we are fetching all and fetch dynamically
-        // Issue #982
-        for file in constants::GENESIS_ASSETS_FILES {
-            let src_path = utils::get_project_path();
-            if let Ok(src_path) = src_path {
-                let src_path = src_path + "/configs/genesis-assets/" + file;
-                utils::copy_from_filesystem(src_path, madara_path.clone() + "/genesis-assets")?;
-            } else {
-                utils::fetch_from_url(
-                    constants::GENESIS_ASSETS_URL.to_string() + file,
-                    madara_path.clone() + "/genesis-assets",
-                )?;
-            }
+    cmd.run_cmd.force_authoring = true;
+    cmd.run_cmd.alice = true;
+
+    // we can't set `--rpc-cors=all`, so it needs to be set manually if we want to connect with external
+    // hosts
+    cmd.run_cmd.rpc_external = true;
+    cmd.run_cmd.rpc_methods = RpcMethods::Unsafe;
+}
+
+fn try_set_testnet(madara_path: &Option<PathBuf>, cmd: &mut ExtendedRunCmd) -> Result<(), String> {
+    // checks if it should retrieve and enable a specific chain-spec
+    let madara_path = get_madara_path_string(madara_path);
+    let local_path = utils::get_project_path();
+
+    if cmd.testnet == Some(Testnet::Sharingan) {
+        if let Ok(ref src_path) = local_path {
+            let src_path = src_path.clone() + "/configs/chain-specs/testnet-sharingan-raw.json";
+            utils::copy_from_filesystem(src_path, madara_path.clone() + "/chain-specs")?;
+            cmd.run_cmd.shared_params.chain = Some(madara_path + "/chain-specs/testnet-sharingan-raw.json");
+        } else {
+            utils::fetch_from_url(
+                constants::SHARINGAN_CHAIN_SPEC_URL.to_string(),
+                madara_path.clone() + "/configs/chain-specs/",
+            )?;
+            cmd.run_cmd.shared_params.chain = Some(madara_path + "/chain-specs/testnet-sharingan-raw.json");
         }
     }
 
-    // TODO confirm with the CI that we are fetching all and fetch dynamically
-    // Issue #982
-    for file in constants::CAIRO_CONTRACTS_FILES {
-        let src_path = utils::get_project_path();
-        if let Ok(src_path) = src_path {
-            let src_path = src_path + "/configs/cairo-contracts/" + file;
-            utils::copy_from_filesystem(src_path, madara_path.clone() + "/cairo-contracts")?;
-        } else {
-            utils::fetch_from_url(
-                constants::CAIRO_CONTRACTS_URL.to_string() + file,
-                madara_path.clone() + "/cairo-contracts",
+    if cmd.run_cmd.shared_params.chain.is_some() {
+        cmd.run_cmd.rpc_external = true;
+        cmd.run_cmd.rpc_methods = RpcMethods::Unsafe;
+    }
+
+    Ok(())
+}
+
+fn set_chain_spec(madara_path: &Option<PathBuf>, cmd: &mut ExtendedRunCmd) -> Result<(), String> {
+    let madara_path = get_madara_path_string(madara_path);
+    let chain_spec_url = cmd
+        .fetch_chain_spec
+        .clone()
+        .expect("`chain_spec_url` expected to be set because the function is called upon verification");
+    utils::fetch_from_url(chain_spec_url.clone(), madara_path.clone() + "/chain-specs")?;
+    let chain_spec =
+        chain_spec_url.split('/').last().expect("Failed to get chain spec file name from `chain_spec_url`");
+    cmd.run_cmd.shared_params.chain = Some(madara_path + "/chain-specs/" + chain_spec);
+
+    Ok(())
+}
+
+fn fetch_madara_configs(madara_path: &Option<PathBuf>, cmd: &SetupCmd) -> Result<(), String> {
+    let madara_path = get_madara_path_string(madara_path);
+    let local_path = utils::get_project_path();
+
+    if let Ok(ref src_path) = local_path {
+        let index_path = src_path.clone() + "/configs/index.json";
+        utils::copy_from_filesystem(index_path, madara_path.clone() + "/configs")?;
+
+        let madara_configs: configs::Configs =
+            serde_json::from_str(&utils::read_file_to_string(madara_path.clone() + "/configs/index.json")?)
+                .expect("Failed to serialize index.json string to json");
+        for asset in madara_configs.genesis_assets {
+            let src_path = src_path.clone() + "/configs/genesis-assets/" + &asset.name;
+            utils::copy_from_filesystem(src_path, madara_path.clone() + "/configs/genesis-assets")?;
+        }
+    } else if let Some(configs_url) = &cmd.fetch_madara_configs {
+        utils::fetch_from_url(configs_url.to_string(), madara_path.clone() + "/configs")?;
+
+        let madara_configs: configs::Configs =
+            serde_json::from_str(&utils::read_file_to_string(madara_path.clone() + "/configs/index.json")?)
+                .expect("Failed to serialize index.json string to json");
+
+        for asset in madara_configs.genesis_assets {
+            configs::fetch_and_validate_file(
+                madara_configs.remote_base_path.clone(),
+                asset,
+                madara_path.clone() + "/configs/genesis-assets/",
             )?;
         }
     }
 
-    if let (Some(chain_spec_url), None) = (cli.run.chain_spec_url.clone(), cli.run.testnet) {
-        utils::fetch_from_url(chain_spec_url, madara_path.clone() + "/chain-specs")?;
-    }
+    Ok(())
+}
 
-    match &cli.subcommand {
-        Some(Subcommand::Key(cmd)) => cmd.run(&cli),
-        Some(Subcommand::BuildSpec(cmd)) => {
+/// Parse and run command line arguments
+pub fn run() -> sc_cli::Result<()> {
+    let mut cli = Cli::from_args();
+
+    match cli.subcommand {
+        Some(Subcommand::Key(ref cmd)) => cmd.run(&cli),
+        Some(Subcommand::BuildSpec(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.sync_run(|config| cmd.run(config.chain_spec, config.network))
         }
-        Some(Subcommand::CheckBlock(cmd)) => {
+        Some(Subcommand::CheckBlock(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, _, import_queue, task_manager, _) = service::new_chain_ops(&mut config)?;
                 Ok((cmd.run(client, import_queue), task_manager))
             })
         }
-        Some(Subcommand::ExportBlocks(cmd)) => {
+        Some(Subcommand::ExportBlocks(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, _, _, task_manager, _) = service::new_chain_ops(&mut config)?;
                 Ok((cmd.run(client, config.database), task_manager))
             })
         }
-        Some(Subcommand::ExportState(cmd)) => {
+        Some(Subcommand::ExportState(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, _, _, task_manager, _) = service::new_chain_ops(&mut config)?;
                 Ok((cmd.run(client, config.chain_spec), task_manager))
             })
         }
-        Some(Subcommand::ImportBlocks(cmd)) => {
+        Some(Subcommand::ImportBlocks(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, _, import_queue, task_manager, _) = service::new_chain_ops(&mut config)?;
                 Ok((cmd.run(client, import_queue), task_manager))
             })
         }
-        Some(Subcommand::PurgeChain(cmd)) => {
+        Some(Subcommand::PurgeChain(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.sync_run(|config| cmd.run(config.database))
         }
-        Some(Subcommand::Revert(cmd)) => {
+        Some(Subcommand::Revert(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, backend, _, task_manager, _) = service::new_chain_ops(&mut config)?;
@@ -169,7 +206,7 @@ pub fn run() -> sc_cli::Result<()> {
                 Ok((cmd.run(client, backend, Some(aux_revert)), task_manager))
             })
         }
-        Some(Subcommand::Benchmark(cmd)) => {
+        Some(Subcommand::Benchmark(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
 
             runner.sync_run(|mut config| {
@@ -234,49 +271,33 @@ pub fn run() -> sc_cli::Result<()> {
         Some(Subcommand::TryRuntime) => Err("TryRuntime wasn't enabled when building the node. You can enable it \
                                              with `--features try-runtime`."
             .into()),
-        Some(Subcommand::ChainInfo(cmd)) => {
+        Some(Subcommand::ChainInfo(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.sync_run(|config| cmd.run::<Block>(&config))
         }
-        None => {
-            // create a reproducible dev environment
-            if cli.run.run_cmd.shared_params.dev {
-                cli.run.run_cmd.shared_params.dev = false;
-                cli.run.run_cmd.shared_params.chain = Some("dev".to_string());
+        Some(Subcommand::Run(ref mut cmd)) => {
+            cmd.run_cmd.shared_params.base_path = cli.madara_path.clone();
+            let madara_path = get_madara_path_string(&cli.madara_path);
 
-                cli.run.run_cmd.force_authoring = true;
-                cli.run.run_cmd.alice = true;
-
-                // we can't set `--rpc-cors=all`, so it needs to be set manually if we want to connect with external
-                // hosts
-                cli.run.run_cmd.rpc_external = true;
-                cli.run.run_cmd.rpc_methods = RpcMethods::Unsafe;
+            // Set the node_key_file for substrate in the case that it was not manually setted
+            if cmd.run_cmd.network_params.node_key_params.node_key_file.is_none() {
+                cmd.run_cmd.network_params.node_key_params.node_key_file =
+                    Some((madara_path.clone() + "/p2p-key.ed25519").into());
             }
 
-            cli.run.run_cmd.network_params.node_key_params.node_key_file =
-                Some((madara_path.clone() + "/p2p-key.ed25519").into());
-
-            if let Some(Testnet::Sharingan) = cli.run.testnet {
-                let src_path = utils::get_project_path();
-                if let Ok(src_path) = src_path {
-                    let src_path = src_path + "/configs/chain-specs/testnet-sharingan-raw.json";
-                    utils::copy_from_filesystem(src_path, madara_path.clone() + "/chain-specs")?;
-                } else {
-                    utils::fetch_from_url(
-                        constants::SHARINGAN_CHAIN_SPEC_URL.to_string(),
-                        madara_path.clone() + "/chain-specs",
-                    )?;
-                }
-
-                cli.run.run_cmd.shared_params.chain =
-                    Some(madara_path.clone() + "/chain-specs/testnet-sharingan-raw.json");
-
-                // This should go apply to all testnets when applying a match pattern
-                cli.run.run_cmd.rpc_external = true;
-                cli.run.run_cmd.rpc_methods = RpcMethods::Unsafe;
+            if cmd.run_cmd.shared_params.dev {
+                set_dev_environment(cmd);
             }
 
-            let da_config: Option<(DaLayer, PathBuf)> = match cli.run.da_layer {
+            if cmd.fetch_chain_spec.is_some() {
+                set_chain_spec(&cli.madara_path, cmd)?;
+            }
+
+            if cmd.testnet.is_some() {
+                try_set_testnet(&cli.madara_path, cmd)?;
+            }
+
+            let da_config: Option<(DaLayer, PathBuf)> = match cmd.da_layer {
                 Some(da_layer) => {
                     let da_path = std::path::PathBuf::from(madara_path.clone() + "/da-config.json");
                     if !da_path.exists() {
@@ -287,15 +308,24 @@ pub fn run() -> sc_cli::Result<()> {
                     Some((da_layer, da_path))
                 }
                 None => {
-                    log::info!("madara initialized w/o da layer");
+                    log::info!("Madara initialized w/o DA layer");
                     None
                 }
             };
 
-            let runner = cli.create_runner(&cli.run.run_cmd)?;
+            // pre assign variables because of cmd mutable borrow
+            let run_cmd: sc_cli::RunCmd = cmd.run_cmd.clone();
+            let sealing = cli.sealing;
+
+            let runner = cli.create_runner(&run_cmd)?;
             runner.run_node_until_exit(|config| async move {
-                service::new_full(config, cli.sealing, da_config).map_err(sc_cli::Error::Service)
+                service::new_full(config, sealing, da_config).map_err(sc_cli::Error::Service)
             })
         }
+        Some(Subcommand::Setup(cmd)) => {
+            fetch_madara_configs(&cli.madara_path, &cmd)?;
+            Ok(())
+        }
+        _ => Err("You need to specify some subcommand. E.g. `madara run`".into()),
     }
 }